check cygwin mismatch earlier
