+++ pre_commit/errors.py
@@ -1,0 +1,2 @@
+class FatalError(RuntimeError):
+    pass
